all functions work!
