+++ ananse/benchmark.py
@@ -1,4 +1,4 @@
-import urllib
+import urllib.request
 import pandas as pd
 import numpy as np
 import re
