+++ ananse/utils.py
@@ -3,15 +3,12 @@ import getpass
 import os
 import pwd
 import shutil
-
-# import subprocess as sp
+import re
+import pandas as pd
 import tempfile
-import warnings
 
 import genomepy.utils
-from pybedtools import BedTool
-
-# import pysam
+from ananse.bed import CombineBedFiles
 
 
 def check_path(arg, error_missing=True):
@@ -37,155 +34,6 @@ def cleanpath(path):
             os.path.expandvars(path)  # expand '$VARIABLES'
         )
     )
-
-
-def shhh_bedtool(func):
-    """
-    Decorator that silences pybedtools RuntimeWarnings such as
-    `line buffering (buffering=1) isn't supported in binary mode`
-    """
-
-    def wrapper(*args, **kwargs):
-        with warnings.catch_warnings():
-            warnings.filterwarnings("ignore", category=RuntimeWarning)
-            func(*args, **kwargs)
-
-    return wrapper
-
-
-@shhh_bedtool
-def bed_sort(bed):
-    """
-    Sort a bed file.
-    """
-    tmpdir = tempfile.mkdtemp(prefix="ANANSE_")
-    try:
-        tmpfile = os.path.join(tmpdir, os.path.basename(bed))
-        BedTool(bed).sort(output=tmpfile)
-        shutil.copy2(tmpfile, bed)
-    finally:
-        shutil.rmtree(tmpdir, ignore_errors=True)
-
-
-@shhh_bedtool
-def bed_merge(list_of_beds, merged_bed):
-    """
-    merge any number of bed files (merges overlapping regions)
-    """
-    bed = BedTool(list_of_beds[0])
-    if list_of_beds[1:]:
-        bed = bed.cat(*list_of_beds[1:])
-    bed.saveas(merged_bed)
-
-
-# @shhh_bedtool
-# def count_reads(bams, peakfile, bed_output):
-#     """
-#     Count bam reads in putative enhancer regions
-#     """
-#     # replace with gimmemotifs.preprocessing.coverage_table()
-#     bed = BedTool(peakfile)
-#     bam_list = bams if isinstance(bams, list) else [bams]
-#     bed.multi_bam_coverage(bams=bam_list, output=bed_output)
-
-
-# def samc(ncore):
-#     """set decent samtools range for samtools functions (1-5 total threads)"""
-#     return max(0, min(ncore - 1, 4))
-#
-#
-# def bam_index(bam, force=True, ncore=1):
-#     if force or not os.path.exists(f"{bam}.bai"):
-#         index_parameters = [f"-@ {samc(ncore)}", bam]
-#         pysam.index(*index_parameters)  # noqa
-#
-#
-# def bam_sort(bam, ncore=1):
-#     tmpdir = tempfile.mkdtemp(prefix="ANANSE_")
-#     try:
-#         sorted_bam = os.path.join(tmpdir, os.path.basename(bam))
-#         sort_parameters = [f"-@ {samc(ncore)}", "-o", sorted_bam, bam]
-#         pysam.sort(*sort_parameters)  # noqa: pysam bug
-#
-#         shutil.copy2(sorted_bam, bam)
-#         bam_index(bam, force=True, ncore=ncore)
-#     finally:
-#         shutil.rmtree(tmpdir, ignore_errors=True)
-#
-#
-# def bam_merge(list_of_bams, merged_bam, ncore=1):
-#     """
-#     merge any number of (sorted) bam files
-#     """
-#     [bam_index(bam, force=False, ncore=ncore) for bam in list_of_bams]
-#     if len(list_of_bams) > 1:
-#         merge_parameters = ["-f", f"-@ {samc(ncore)}", merged_bam] + list_of_bams
-#         pysam.merge(*merge_parameters)  # noqa: pysam bug
-#         bam_index(merged_bam)
-#     else:
-#         # os.symlink() doesn't work with multi_bam_coverage()
-#         bam = list_of_bams[0]
-#         shutil.copy2(bam, merged_bam)
-#         shutil.copy2(f"{bam}.bai", f"{merged_bam}.bai")
-
-
-# def mosdepth(bed, bam, bed_output, ncore=1):
-#     """
-#     Count (median per base overlap of) bam reads in putative enhancer regions
-#     """
-#     ncore = min(4, ncore)
-#     tmpdir = tempfile.mkdtemp(prefix="ANANSE_")
-#     try:
-#         prefix = os.path.join(tmpdir, "bam_coverage")
-#         cmd = f"mosdepth -nxm -t {ncore} -b {bed} {prefix} {bam}"
-#         sp.check_call(cmd, shell=True)
-#
-#         tmp_bed_output = f"{prefix}.regions.bed"
-#         cmd = f"gunzip -f {tmp_bed_output}.gz"
-#         sp.check_call(cmd, shell=True)
-#
-#         shutil.copy2(tmp_bed_output, bed_output)
-#     finally:
-#         shutil.rmtree(tmpdir, ignore_errors=True)
-
-
-# def bed_sum_coverages(multi_bam_coverage, sum_bam_coverage):
-#     """
-#     MultiBamCov returns a BED3+n with one column per bam.
-#     This function sums up all bam counts and returns a BED3+1.
-#     """
-#     bed = pd.read_csv(multi_bam_coverage, header=None, sep="\t")
-#     columns = bed.shape[1]
-#     if columns != 4:
-#         bed3 = bed.iloc[:, :3]
-#         scores = bed.iloc[:, 3:].sum(axis=1)
-#         bed = pd.concat([bed3, scores], axis=1)
-#     bed.to_csv(sum_bam_coverage, sep="\t", header=False, index=False)
-
-
-# def non_empty_files(files, error_msg, size_threshold=10, verbose=True):
-#     """Check list of files for content
-#
-#     Args:
-#         files: list of filepaths
-#         error_msg: message for all empty files
-#         size_threshold: minimum size to be considered non-empty
-#         verbose: return warnings?
-#
-#     Returns:
-#         list of non-empty files
-#     """
-#     ret_files = []
-#     for file in files:
-#         if os.path.getsize(file) > size_threshold:
-#             ret_files.append(file)
-#         elif verbose:
-#             logger.warning(f"Empty file: '{os.path.basename(file)}'")
-#
-#     if not ret_files:
-#         logger.exception(error_msg)
-#         exit(1)
-#     return ret_files
 
 
 def mytmpdir():
@@ -247,21 +95,92 @@ def check_input_factors(factors):
     if factors is None:
         return
 
+    fname = None
     # if factors is a string, assume it's a filename
     if isinstance(factors, str):
         fname = factors
-
-    # if factors is a list of 1, and it exists, assume it's a filename
+    # if factors is a list of 1, assume it's a filename
     elif isinstance(factors, list) and len(factors) == 1:
         fname = factors[0]
 
-    # It's a list with more than one value, assuming it's a list of TF names.
-    else:
+    # if the filename exists, use that.
+    # else, it's a list, assuming it's a list of TF names.
+    if fname is None or not os.path.exists(cleanpath(fname)):
         return factors
 
-    fname = cleanpath(fname)
-    if not os.path.exists(fname):
-        raise ValueError(f"Factors file '{factors}' does not exist")
-
-    factors = [line.strip() for line in open(fname)]
+    factors = [line.strip() for line in open(cleanpath(fname))]
     return factors
+
+
+def check_input_regions(
+    regionfiles, genome=None, outdir=".", verbose=True, force=False
+):
+    """
+    Check argument {regionfiles}.
+    If its a list of multiple files, assume they are BED files (e.g. narrowPeak) and merge them.
+    Then convert the (merged) regionfile to regions (e.g. chr1:100-200).
+
+    Parameters
+    ----------
+    regionfiles: None or str or list
+        one or more BED files, or up to one region file
+    genome: str, optional
+        fasta file. Only needed if regionfiles need to be merged
+    outdir: str, optional
+        path. Only needed if regionfiles need to be merged
+    verbose: bool, optional
+        Narrate merging? Only needed if regionfiles need to be merged
+    force: bool, optional
+        Overwrite merged bed? Only needed if regionfiles need to be merged
+
+    Returns
+    -------
+    list
+        unique regions
+    """
+    if regionfiles is None:
+        # Keep regions to None, use reference regions.
+        return
+
+    # accept a file or list of files
+    if isinstance(regionfiles, str):
+        infile = regionfiles
+    elif len(regionfiles) == 1:
+        infile = regionfiles[0]
+    else:
+        # merge region files, assumed to be all BED
+        peak_width = 200
+        cbed = CombineBedFiles(genome=genome, peakfiles=regionfiles, verbose=verbose)
+        combined_bed = os.path.join(outdir, "regions_combined.bed")
+        cbed.run(outfile=combined_bed, width=peak_width, force=force)
+        infile = combined_bed
+
+    df = pd.read_table(infile, header=None, sep="\t", comment="#", dtype=str)
+    assert df.shape[0] > 1, "regions file must have more than 1 region!"
+
+    test = str(df.at[1, 0])  # skip potential header line
+    if bool(re.match(r"^.*:\d+-\d+$", test)):
+        # it's a regions list
+        # or it's a Seq2science counts table
+        regions = df.iloc[:, 0].tolist()
+
+    elif df.shape[1] >= 3:
+        # it's a BED file
+        regions = (
+            # For Ensembl genome names, make sure it's a string
+            df.iloc[:, 0].astype(str)
+            + ":"
+            + df.iloc[:, 1].astype(str)
+            + "-"
+            + df.iloc[:, 2].astype(str)
+        ).tolist()
+
+    else:
+        raise TypeError("Cannot identify regions file(s) type.")
+
+    # remove the header, if any.
+    header = str(regions[0])
+    if not bool(re.match(r"^.*:\d+-\d+$", header)):
+        regions = regions[1:]
+
+    return list(set(regions))
