+++ docs/command-line_reference.md
@@ -165,7 +165,7 @@ If you specify a file, e.g. `mymotifs.pfm`, the command expects `mymotifs.motif2
 #### Full options
 
 ```shell
-usage: ananse [-h] <command> [options] binding [-A BAM [BAM ...]] [-H BAM [BAM ...]] [-g NAME] [-r FILE [FILE ...]] [-p FILE] [-o DIR] [-R DIR] [--pfmscorefile FILE] [-f [TF ...]] [--jaccard-cutoff FLOAT] [-n INT] [-h]
+usage: ananse [-h] <command> [options] binding [-A BAM [BAM ...]] [-H BAM [BAM ...]] [-g NAME] [-r FILE [FILE ...]] [-p FILE] [-o DIR] [-R DIR] [--pfmscorefile FILE] [-t [TF ...]] [--jaccard-cutoff FLOAT] [-n INT] [-h]
 
 Required arguments:
   -A BAM [BAM ...], --atac-bams BAM [BAM ...]
@@ -186,8 +186,8 @@ Optional arguments:
   -R DIR, --reference DIR
                         Path to reference data directory
   --pfmscorefile FILE   Use precomputed gimmemotifs scores (gimme scan -Tz --gc -g GENOME REGIONS > SCAN.tsv)
-  -f [TF ...], --factors [TF ...]
-                        Transcription factors to use. Either a space-separated list or a file with one TF per line
+  -t [TF ...], --tfs [TF ...]
+                        Filter Transcription Factors to use (default: all in motif2factors.txt). Either a space-separated list or a file with one TF per line
   --jaccard-cutoff FLOAT
                         TFs with a jaccard motif similarity >= the cutoff can be used as backup model. 0: any similarity, 1: perfect similarity (default is 0)
   -n INT, --ncore INT   Number of cores to use.
@@ -229,14 +229,13 @@ If gene symbols are converted from transcripts/gene ids to gene names, expressio
 #### Full options
 
 ```shell
-usage: ananse [-h] <command> [options] network -b FILE -e FILE [FILE ...] [-g NAME] [-a BED] [-o FILE] [-f] [--include-promoter] [--include-enhancer] [-n INT] [-h]
+usage: ananse [-h] <command> [options] network -b FILE -e FILE [FILE ...] [-g NAME] [-a BED] [-o FILE] [-t [TF ...]] [-r FILE] [-f] [--include-promoter] [--include-enhancer] [-n INT] [-h]
 
 required arguments:
   -b FILE, --binding FILE
                         TF binding prediction file (from ANANSE binding).
   -e FILE [FILE ...], --expression FILE [FILE ...]
-                        Gene expression files with gene name (HGNC symbol in case of human) as first column and a column name 'TPM' (case insensitive). This file can be created by summarizing transcript-level TPMs (the
-                        quant.sf from salmon or the abundances.tsv from kallisto) to gene-level TPMs with tximeta.
+                        Gene expression files with gene name (HGNC symbol in case of human) as first column and a column name 'TPM' (case insensitive). This file can be created by summarizing transcript-level TPMs (the quant.sf from salmon or the abundances.tsv from kallisto) to gene-level TPMs with tximeta.
 
 Required arguments (optional for hg38):
   -g NAME, --genome NAME
@@ -247,6 +246,10 @@ Required arguments (optional for hg38):
 optional arguments:
   -o FILE, --outfile FILE
                         Name of the output network file (default: ./ANANSE_network.tsv)
+  -t [TF ...], --tfs [TF ...]
+                        Filter Transcription Factors to use (default: all in binding.h5). Either a space-separated list or a file with one TF per line
+  -r FILE, --regions FILE
+                        Filter regions to use (default: all in binding.h5). One region/BED format file.
   -f, --full-output     Export the full GRN output to the output file
   --include-promoter, --exclude-promoter
                         Include or exclude promoter peaks (<= TSS +/- 2kb) in network inference. By default promoter peaks are included.
@@ -335,16 +338,16 @@ $ ananse plot -i results/influence.tsv \
 #### Full options
 
 ```shell
-usage: ananse [-h] <command> [options] plot -i FILE [-o DIR] [--diff-network GRN_FILE] [--edge-info EDGE_INFO] [--edge-min EDGE_MIN] [--node-placement NETWORK_ALGORITHM] [--n-tfs N_TFS] [-c CMAP] [-f] [-h]
+usage: ananse [-h] <command> [options] plot -i FILE [-o DIR] [-d--diff-network GRN_FILE] [--edge-info EDGE_INFO] [--edge-min EDGE_MIN] [--node-placement NETWORK_ALGORITHM] [--n-tfs N_TFS] [-c CMAP] [-f] [-t FTYPE] [-h]
 
 required arguments:
   -i FILE, --influence-file FILE
                         ANANSE influence file
 
 optional arguments:
-  -o DIR, --outdir DIR  Directory where you wish to store the output (default: ./ANANSE_binding)
-  --diff-network GRN_FILE
-                        Differential network file generated by ANANSE influence
+  -o DIR, --outdir DIR  Directory where you wish to store the output (default: ./ANANSE_plot)
+  -d--diff-network GRN_FILE
+                        ANANSE influence diffnetwork file (automatically generated)
   --edge-info EDGE_INFO
                         Column to use for edges of GRN, default: 'weight'. When full_output is specified, options are 'wb_diff' ,'tf_act_diff', 'tf_expr_diff', 'tg_expr_diff'
   --edge-min EDGE_MIN   Minimum value for an edge to be included in the GRN image
@@ -353,6 +356,8 @@ optional arguments:
   --n-tfs N_TFS         Amount of TFs to plot in the GRN, default is top 20 differential TFs
   -c CMAP, --cmap CMAP  matlotlib colour library
   -f, --full-output     Select if the diffnetwork is a full output file
+  -t FTYPE, --type FTYPE
+                        Specify the output filetype (default: pdf)
   -h, --help            show this help message and exit
 ```
