+++ scripts/ananse
@@ -120,10 +120,11 @@ if __name__ == "__main__":
         default=None,
     )
     group.add_argument(
-        "-f",
-        "--factors",
+        "-t",
+        "--tfs",
         dest="factors",
-        help="Transcription factors to use. Either a space-separated list or a file with one TF per line",
+        help="Filter Transcription Factors to use (default: all in motif2factors.txt). "
+             "Either a space-separated list or a file with one TF per line",
         metavar="TF",
         default=None,
         nargs="*",
@@ -205,6 +206,24 @@ if __name__ == "__main__":
         help="Name of the output network file (default: ./ANANSE_network.tsv)",
         metavar="FILE",
         default="./ANANSE_network.tsv",
+    )
+    group.add_argument(
+        "-t",
+        "--tfs",
+        dest="factors",
+        help="Filter Transcription Factors to use (default: all in binding.h5). "
+             "Either a space-separated list or a file with one TF per line",
+        metavar="TF",
+        default=None,
+        nargs="*",
+    )
+    group.add_argument(
+        "-r",
+        "--regions",
+        dest="regions",
+        help="Filter regions to use (default: all in binding.h5). One region/BED format file.",
+        metavar="FILE",
+        default=None,
     )
     group.add_argument(
         "-f",
@@ -430,19 +449,19 @@ if __name__ == "__main__":
         default=None,
         required=True,
     )
-
     group = p.add_argument_group("optional arguments")
     group.add_argument(
         "-o",
         "--outdir",
         dest="outdir",
-        help="Directory where you wish to store the output (default: ./ANANSE_binding)",
+        help="Directory where you wish to store the output (default: ./ANANSE_plot)",
         metavar="DIR",
-        default="./ANANSE_binding",
+        default="./ANANSE_plot",
     )
     group.add_argument(
+        "-d"
         "--diff-network",
-        help="Differential network file generated by ANANSE influence",
+        help="ANANSE influence diffnetwork file (automatically generated)",
         dest="GRN_file",
         default="",
     )
@@ -490,6 +509,13 @@ if __name__ == "__main__":
         help="Select if the diffnetwork is a full output file",
         action="store_true",
         default=False,
+    )
+    group.add_argument(
+        "-t",
+        "--type",
+        dest="ftype",
+        help="Specify the output filetype (default: pdf)",
+        default="pdf",
     )
     group.add_argument(
         "-h", "--help", action="help", help="show this help message and exit"
